Fix up XCode projects a bit.
Change the location of the plug-ins in the Mulberry project to be based
on BUILT_PRODUCTS_DIR instead of instead of SOURCE_ROOT.  This makes
the project work regardless of the setting of the "Build Location"
XCode preference.  In order to make this work for both Debug and
Release builds of the main project, Release builds were added to each
plug-in project.  Some other changes were made by XCode 4 on its own.