Change the build products location for plug-ins to ../../MacOS/build
This is needed to make the build work when the XCode "Build Location"
preference is set to "Location Specified by Targets".  More importantly
the command line tool ignores this preference and always acts as if it
is on.  As a beneficial side effect, all the build directories in the
Plug-ins directory disappear.